requirements.txt
kagglehub
pandas
scikit-learn
numpy
fastapi
uvicornpython-multipart